- Modified the Floating Action Button with a Floating Action Menu which gives the option to search for devices or allows you to manually add a device.
I think the icons are backwards but the functionallity is there.  The idea is you click the "+" button and you get 2 options.  Manual, or Auto.

When you've created your class that get's the devices with JSON automatically and adds it to the database, have it call that activity from the autoAddFab.setOnClickListener() event.  We can discuss tonight.